requirements.txt
Django
djangorestframework
gunicorn
psycopg2
dj_database_url
